fix to prevent tensorflow reinstallation in setup
Removed deepcell-tf from the requirements.txt file.
Added deepcell-tf installation to the Dockerfile, in which deepcell-tf is cloned, and tensorflow is removed from its requirements.txt before installation. Previously, tensorflow would get reinstalled in the container during the installation of deepcell-tf, resulting in a container that has a tensorflow version that doesn't recognize the GPU.